V1.6
Some clean ups.
Proper support of the Disabled attribute.
Support for more scene and notification triggers
Use of dkjson
Tested for openLuup. Need to install package lua-bitop (sudo apt-get install lua-bitop)
For UI7. In Vacation mode the command TC= is used, in Away mode TT=. This way the thermostat program can warm up the house for the evening when away for the day.
Note: not all thermostats keep the setpoint with TC=, some have the program overwrite anyway. An update for that situation is in the planning.
Optimized for UI7 and ALTUI.
Added 60 seconds polling of Home / Away status instead of monitoring an alarm pannel status. When Away/Vacation ECO options are used, when Home restored.
Use of otg. variable container to reduce number of variabled (Vera can handle up to 60, but less is better)